pyproject.toml
 [project]
name = "procode_framework"
version = "0.1.0"
description = "Pro-code agent skeleton (A2A-ready)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.4",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "rich>=13.7.0",
    "sse-starlette>=2.3.5",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
